# docs/swagger.yaml
basePath: /
definitions:
...
  protocol.CallbackResponse:
    properties:
      accessToken:
        type: string
      refreshToken:
        type: string
    type: object
...
  protocol.Comment:
    properties:
      commentID:
        type: integer
      content:
        type: string
      createdAt:
        type: string
      likes:
        type: integer
      replyTo:
        type: integer
      userID:
        type: integer
    type: object
...
  protocol.CreateArticleCommentBody:
    properties:
      articleID:
        type: integer
      content:
        maxLength: 300
        minLength: 1
        type: string
      replyTo:
        type: integer
    required:
    - articleID
    - content
    type: object
  protocol.CreateArticleCommentResponse:
    properties:
      comment:
        $ref: '#/definitions/protocol.Comment'
    type: object
...
  protocol.DeleteCommentResponse:
    type: object
...
  protocol.HTTPResponse:
    properties:
      data: {}
      error:
        type: string
    type: object
...
  protocol.ListArticleCommentsResponse:
    properties:
      comments:
        items:
          $ref: '#/definitions/protocol.Comment'
        type: array
      pageInfo:
        $ref: '#/definitions/protocol.PageInfo'
    type: object
...
  protocol.ListChildrenCommentsResponse:
    properties:
      comments:
        items:
          $ref: '#/definitions/protocol.Comment'
        type: array
      pageInfo:
        $ref: '#/definitions/protocol.PageInfo'
    type: object
...
  protocol.LoginResponse:
    properties:
      redirectURL:
        type: string
    type: object
  protocol.PageInfo:
    properties:
      page:
        type: integer
      pageSize:
        type: integer
      total:
        type: integer
    type: object
...
host: 9.134.115.68:8170
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: Aris-blog API
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Aris-blog
  version: "1.0"
paths:
...
  /v1/comment:
    post:
      consumes:
      - application/json
      description: 创建文章评论
      parameters:
      - description: 创建文章评论请求
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/protocol.CreateArticleCommentBody'
      produces:
      - application/json
      responses:
        "200":
          description: 创建文章评论响应
          schema:
            allOf:
            - $ref: '#/definitions/protocol.HTTPResponse'
            - properties:
                data:
                  $ref: '#/definitions/protocol.CreateArticleCommentResponse'
                error:
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/protocol.HTTPResponse'
            - properties:
                data:
                  type: object
                error:
                  type: string
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/protocol.HTTPResponse'
            - properties:
                data:
                  type: object
                error:
                  type: string
              type: object
        "403":
          description: Forbidden
          schema:
            allOf:
            - $ref: '#/definitions/protocol.HTTPResponse'
            - properties:
                data:
                  type: object
                error:
                  type: string
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/protocol.HTTPResponse'
            - properties:
                data:
                  type: object
                error:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 创建文章评论
      tags:
      - comment
  /v1/comment/{commentID}:
    delete:
      consumes:
      - application/json
      description: 删除评论
      parameters:
      - in: path
        minimum: 1
        name: commentID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 删除评论响应
          schema:
            allOf:
            - $ref: '#/definitions/protocol.HTTPResponse'
            - properties:
                data:
                  $ref: '#/definitions/protocol.DeleteCommentResponse'
                error:
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/protocol.HTTPResponse'
            - properties:
                data:
                  type: object
                error:
                  type: string
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/protocol.HTTPResponse'
            - properties:
                data:
                  type: object
                error:
                  type: string
              type: object
        "403":
          description: Forbidden
          schema:
            allOf:
            - $ref: '#/definitions/protocol.HTTPResponse'
            - properties:
                data:
                  type: object
                error:
                  type: string
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/protocol.HTTPResponse'
            - properties:
                data:
                  type: object
                error:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 删除评论
      tags:
      - comment
  /v1/comment/{commentID}/subComments:
    get:
      consumes:
      - application/json
      description: 列出子评论
      parameters:
      - in: path
        minimum: 1
        name: commentID
        required: true
        type: integer
      - in: query
        minimum: 1
        name: page
        required: true
        type: integer
      - in: query
        maximum: 50
        minimum: 1
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 列出子评论响应
          schema:
            allOf:
            - $ref: '#/definitions/protocol.HTTPResponse'
            - properties:
                data:
                  $ref: '#/definitions/protocol.ListChildrenCommentsResponse'
                error:
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/protocol.HTTPResponse'
            - properties:
                data:
                  type: object
                error:
                  type: string
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/protocol.HTTPResponse'
            - properties:
                data:
                  type: object
                error:
                  type: string
              type: object
        "403":
          description: Forbidden
          schema:
            allOf:
            - $ref: '#/definitions/protocol.HTTPResponse'
            - properties:
                data:
                  type: object
                error:
                  type: string
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/protocol.HTTPResponse'
            - properties:
                data:
                  type: object
                error:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 列出子评论
      tags:
      - comment
  /v1/comment/article/{articleID}/list:
    get:
      consumes:
      - application/json
      description: 列出文章评论
      parameters:
      - in: path
        name: articleID
        required: true
        type: integer
      - in: query
        minimum: 1
        name: page
        required: true
        type: integer
      - in: query
        maximum: 50
        minimum: 1
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 列出文章评论响应
          schema:
            allOf:
            - $ref: '#/definitions/protocol.HTTPResponse'
            - properties:
                data:
                  $ref: '#/definitions/protocol.ListArticleCommentsResponse'
                error:
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/protocol.HTTPResponse'
            - properties:
                data:
                  type: object
                error:
                  type: string
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/protocol.HTTPResponse'
            - properties:
                data:
                  type: object
                error:
                  type: string
              type: object
        "403":
          description: Forbidden
          schema:
            allOf:
            - $ref: '#/definitions/protocol.HTTPResponse'
            - properties:
                data:
                  type: object
                error:
                  type: string
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/protocol.HTTPResponse'
            - properties:
                data:
                  type: object
                error:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 列出文章评论
      tags:
      - comment
  /v1/oauth2/github/callback:
    get:
      consumes:
      - application/json
      description: Github OAuth2回调请求,验证code和state
      parameters:
      - description: 授权码
        in: query
        name: code
        required: true
        type: string
      - description: 状态码
        in: query
        name: state
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/protocol.HTTPResponse'
            - properties:
                data:
                  $ref: '#/definitions/protocol.CallbackResponse'
                error:
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/protocol.HTTPResponse'
            - properties:
                data:
                  type: object
                error:
                  type: string
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/protocol.HTTPResponse'
            - properties:
                data:
                  type: object
                error:
                  type: string
              type: object
        "403":
          description: Forbidden
          schema:
            allOf:
            - $ref: '#/definitions/protocol.HTTPResponse'
            - properties:
                data:
                  type: object
                error:
                  type: string
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/protocol.HTTPResponse'
            - properties:
                data:
                  type: object
                error:
                  type: string
              type: object
      summary: Github OAuth2回调
      tags:
      - oauth2
  /v1/oauth2/github/login:
    get:
      consumes:
      - application/json
      description: Github OAuth2登录请求,返回重定向URL
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/protocol.HTTPResponse'
            - properties:
                data:
                  $ref: '#/definitions/protocol.LoginResponse'
                error:
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/protocol.HTTPResponse'
            - properties:
                data:
                  type: object
                error:
                  type: string
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/protocol.HTTPResponse'
            - properties:
                data:
                  type: object
                error:
                  type: string
              type: object
        "403":
          description: Forbidden
          schema:
            allOf:
            - $ref: '#/definitions/protocol.HTTPResponse'
            - properties:
                data:
                  type: object
                error:
                  type: string
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/protocol.HTTPResponse'
            - properties:
                data:
                  type: object
                error:
                  type: string
              type: object
      summary: Github OAuth2登录
      tags:
      - oauth2
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"